backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.8
frozenlist==1.5.0
gotrue==2.11.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
postgrest==0.19.3
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.3.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
